implementacion de moviemiento de personajes
