+++ index.html
@@ -6,6 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Juego de peleas</title>
     <link rel="stylesheet" href="style.css">
+    <script src="./js/main.js" type="module"></script>
   </head>
 
   <body>
@@ -13,15 +14,13 @@
       <h2>Fight!</h2>
       <h3>Cada ataque tiene daño/ataque entre 5 - 10</h3>
     </div>
-
-    <script src="./js/solucion_ejercicio_2.js"></script>
     <div class="vida-jugadores">
       <div class="persona">
         <div>
           <h4>Hero</h4>
         </div>
         <div class="progress">
-          <div class="barra-vida-hero" id="hero" style="width: 100%;"></div>
+          <div class="barra-vida-hero" id="heroe-vida" style="width: 100%;"></div>
         </div>
     </div>
     <div class="persona">
@@ -29,7 +28,7 @@
         <h4>Limo</h4>
       </div>
       <div class="progress">
-        <div class="barra-vida-enemy" id="enemy" style="width: 100%;"></div>
+        <div class="barra-vida-enemy" id="enemigo-vida" style="width: 100%;"></div>
       </div>
   </div>
     </div>
