--- js/controlador_personajes.js
@@ -1,33 +1,0 @@
-document.addEventListener("keydown", function(event){
-    if(event === "a") {
-
-    }
-    if(event === "d"){
-
-    }
-    if(event === "w"){
-
-    }
-    if(event === "s") {
-
-    }
-});
-
-document.addEventListener("keydown", function(event) {
-    if (event.key === "ArrowLeft") {
-        console.log("Izquierda");
-    }
-    if (event.key === "ArrowRight") {
-        console.log("Derecha");
-    }
-    if (event.key === "ArrowUp") {
-        console.log("Arriba");
-    }
-    if (event.key === "ArrowDown") {
-        console.log("Abajo");
-    }
-});
-
-function moverPersonajeArriba() {
-
-}
