+++ style.css
@@ -10,6 +10,7 @@ html {
 .progress {
     margin-top: 20px;
     margin-right: 150px;
+    margin-left: 20px;
     width: 300px;
     height: 20px;
     background-color: #ddd;
